Extract zkchat routes in own controller
