+++ src/services/http/controllers/Tags.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from 'express';
-import { makeResponse } from '../utils';
+
 import { Controller } from './interface';
+import { makeResponse } from '../utils';
 
 export class TagsController extends Controller {
   constructor() {
