+++ src/services/http/controllers/index.ts
@@ -1,3 +1,4 @@
 export * from './Posts';
 export * from './Tags';
 export * from './Users';
+export * from './ZkChat';
