+++ src/services/http/index.ts
@@ -1,12 +1,11 @@
 import { GenericService } from '../../util/svc';
 import express, { Express, NextFunction, Request, Response } from 'express';
 import bodyParser from 'body-parser';
-import cors, { CorsOptions } from 'cors';
+import cors from 'cors';
 import http from 'http';
 import config from '../../util/config';
 import logger from '../../util/logger';
 import path from 'path';
-import Web3 from 'web3';
 import { getLinkPreview } from 'link-preview-js';
 import queryString from 'querystring';
 import session from 'express-session';
@@ -17,7 +16,6 @@ import {
   accessToken,
   createHeader,
   getBotometerScore,
-  getReplies,
   getUser,
   requestToken,
   showStatus,
@@ -26,23 +24,14 @@ import {
   verifyCredential,
 } from '../../util/twitter';
 import { verifySignatureP256 } from '../../util/crypto';
-import { parseMessageId, PostMessageSubType } from '../../util/message';
 import fs from 'fs';
 import { getFilesFromPath } from 'web3.storage';
 import { UploadModel } from '../../models/uploads';
 import { genExternalNullifier, Semaphore, SemaphoreFullProof } from '@zk-kit/protocols';
 import vKey from '../../../static/verification_key.json';
-import merkleRoot from '../../models/merkle_root';
 import { sequelize } from '../../util/sequelize';
 import crypto from 'crypto';
-import {
-  addConnection,
-  addTopic,
-  keepAlive,
-  publishTopic,
-  removeConnection,
-  SSEType,
-} from '../../util/sse';
+import { addConnection, addTopic, keepAlive, removeConnection } from '../../util/sse';
 import { makeResponse, upload } from './utils';
 import { corsOptions, JWT_SECRET, maxFileSize, maxPerUserSize } from './constants';
 
@@ -52,7 +41,6 @@ const SequelizeStore = require('connect-session-sequelize')(session.Store);
 export default class HttpService extends GenericService {
   app: Express;
   httpServer: any;
-  merkleRoot?: ReturnType<typeof merkleRoot>;
 
   constructor() {
     super();
@@ -162,149 +150,6 @@ export default class HttpService extends GenericService {
     };
   }
 
-  handleGetChatUsers = async (req: Request, res: Response) => {
-    const limit = req.query.limit && Number(req.query.limit);
-    const offset = req.query.offset && Number(req.query.offset);
-    const users = await this.call('zkchat', 'getAllUsers', offset, limit);
-    res.send(makeResponse(users));
-  };
-
-  handlePostChatMessage = async (req: Request, res: Response) => {
-    const { messageId, type, timestamp, sender, receiver, ciphertext, rln, semaphore } = req.body;
-    const signature = req.header('X-SIGNED-ADDRESS');
-    const userDB = await this.call('db', 'getUsers');
-
-    if (!sender.address && !sender.hash && !sender.ecdh) throw new Error('invalid sender');
-    if (!receiver.address && !receiver.ecdh) throw new Error('invalid receiver');
-
-    if (rln) {
-      if (!sender.hash) throw new Error('invalid request object');
-
-      const isEpochCurrent = await this.call('zkchat', 'isEpochCurrent', rln.epoch);
-      const verified = await this.call('zkchat', 'verifyRLNProof', rln);
-      const root = '0x' + BigInt(rln.publicSignals.merkleRoot).toString(16);
-      const group = await this.call('merkle', 'getGroupByRoot', root);
-
-      if (!isEpochCurrent) throw new Error('outdated message');
-      if (!verified) throw new Error('invalid rln proof');
-      if (!group) throw new Error('invalid merkle root');
-
-      rln.group_id = group;
-
-      const share = {
-        nullifier: rln.publicSignals.internalNullifier,
-        epoch: rln.publicSignals.epoch,
-        y_share: rln.publicSignals.yShare,
-        x_share: rln.x_share,
-      };
-
-      const { isSpam, isDuplicate } = await this.call('zkchat', 'checkShare', share);
-
-      if (isDuplicate) {
-        throw new Error('duplicate message');
-      }
-
-      if (isSpam) {
-        res.status(429).send('too many requests');
-        return;
-      }
-
-      await this.call('zkchat', 'insertShare', share);
-      await this.merkleRoot?.addRoot(root, group);
-    } else if (semaphore) {
-      const verified = await this.call('zkchat', 'verifySemaphoreProof', semaphore);
-      const root = '0x' + BigInt(semaphore.publicSignals.merkleRoot).toString(16);
-      const group = await this.call('merkle', 'getGroupByRoot', root);
-      if (!verified) throw new Error('invalid proof');
-      if (!group) throw new Error('invalid merkle root');
-      await this.merkleRoot?.addRoot(root, group);
-    } else if (signature) {
-      const [sig, address] = signature.split('.');
-      const user = await userDB.findOneByName(address);
-
-      if (user?.pubkey) {
-        if (!verifySignatureP256(user.pubkey, address, sig)) {
-          res.status(403).send(makeResponse('unauthorized', true));
-          return;
-        }
-      }
-    } else {
-      res.status(403).send(makeResponse('unauthorized', true));
-      return;
-    }
-
-    const data = await this.call('zkchat', 'addChatMessage', {
-      messageId,
-      type,
-      timestamp: new Date(timestamp),
-      sender,
-      receiver,
-      ciphertext,
-      rln,
-    });
-
-    // await
-    publishTopic(`ecdh:${data.sender_pubkey}`, {
-      type: SSEType.NEW_CHAT_MESSAGE,
-      message: data,
-    });
-    // await?
-    publishTopic(`ecdh:${data.receiver_pubkey}`, {
-      type: SSEType.NEW_CHAT_MESSAGE,
-      message: data,
-    });
-    res.send(makeResponse(data));
-  };
-
-  handleGetDirectMessage = async (req: Request, res: Response) => {
-    const { sender, receiver } = req.params;
-    const limit = req.query.limit && Number(req.query.limit);
-    const offset = req.query.offset && Number(req.query.offset);
-    const data = await this.call('zkchat', 'getDirectMessages', sender, receiver, offset, limit);
-    res.send(makeResponse(data));
-  };
-
-  handleGetDirectChats = async (req: Request, res: Response) => {
-    const { pubkey } = req.params;
-
-    const values = await sequelize.query(
-      // prettier-ignore
-      `
-          SELECT distinct zkc.receiver_pubkey as pubkey, zkc.receiver_address as address, null as group_id
-          FROM zkchat_chats zkc
-          WHERE zkc.receiver_pubkey IN (SELECT distinct receiver_pubkey FROM zkchat_chats WHERE sender_pubkey = :pubkey)
-          UNION
-          SELECT distinct zkc.sender_pubkey as pubkey, zkc.sender_address as address, mr.group_id
-          FROM zkchat_chats zkc
-                   LEFT JOIN merkle_roots mr on mr.root_hash = zkc.rln_root
-          WHERE zkc.sender_pubkey IN (SELECT distinct sender_pubkey FROM zkchat_chats WHERE receiver_pubkey = :pubkey);
-      `,
-      {
-        type: QueryTypes.SELECT,
-        replacements: {
-          pubkey,
-        },
-      }
-    );
-
-    const data = values.map((val: any) => ({
-      type: 'DIRECT',
-      receiver: val.address,
-      receiverECDH: val.pubkey,
-      senderECDH: pubkey,
-      group: val.group_id,
-    }));
-
-    res.send(makeResponse(data));
-  };
-
-  handleSearchChats = async (req: Request, res: Response) => {
-    const { query } = req.params;
-    const { sender } = req.query;
-    const data = await this.call('zkchat', 'searchChats', query || '', sender);
-    res.send(makeResponse(data));
-  };
-
   handleGetProofs = async (req: Request, res: Response) => {
     const { idCommitment } = req.params;
     const { group = '', proofType = '' } = req.query;
@@ -390,7 +235,7 @@ export default class HttpService extends GenericService {
   };
 
   initControllers() {
-    ['users', 'posts', 'tags'].forEach(controller => {
+    ['users', 'posts', 'tags', 'zkChat'].forEach(controller => {
       this.app.use('/v1', this.get(`${controller}Controller`, 'router'));
     });
   }
@@ -403,19 +248,6 @@ export default class HttpService extends GenericService {
         res.send(makeResponse('ok'));
       })
     );
-
-    this.app.get('/v1/zkchat/users', this.wrapHandler(this.handleGetChatUsers));
-    this.app.post(
-      '/v1/zkchat/chat-messages',
-      jsonParser,
-      this.wrapHandler(this.handlePostChatMessage)
-    );
-    this.app.get(
-      '/v1/zkchat/chat-messages/dm/:sender/:receiver',
-      this.wrapHandler(this.handleGetDirectMessage)
-    );
-    this.app.get('/v1/zkchat/chats/dm/:pubkey', this.wrapHandler(this.handleGetDirectChats));
-    this.app.get('/v1/zkchat/chats/search/:query?', this.wrapHandler(this.handleSearchChats));
 
     this.app.get('/v1/proofs/:idCommitment', this.wrapHandler(this.handleGetProofs));
     this.app.get('/v1/group_members/:group', this.wrapHandler(this.handleGetMembers));
@@ -838,7 +670,6 @@ export default class HttpService extends GenericService {
   }
 
   async start() {
-    this.merkleRoot = merkleRoot(sequelize);
     const httpServer = http.createServer(this.app);
 
     this.app.set('trust proxy', 1);
